refactor: Component 내 폴더구조 변경 - 특정 도메인에 종속되는 컴포넌트 (ShoppingCart) - 공통으로 사용할 수 있는 컴포넌트 (common) - 에러 관련 컴포넌트 (Error)
